+++ Block/Adminhtml/Campaign/Edit/BackButton.php
@@ -1,0 +1,19 @@
+<?php
+declare(strict_types=1);
+
+namespace Ordo\Automation\Block\Adminhtml\Campaign\Edit;
+
+use Magento\Ui\Component\Control\Container\ToolbarButtonInterface;
+
+class BackButton extends GenericButton implements ToolbarButtonInterface
+{
+    public function getButtonData(): array
+    {
+        return [
+            'label' => __('Back'),
+            'on_click' => sprintf("location.href = '%s';", $this->getUrl('*/*/')),
+            'class' => 'back',
+            'sort_order' => 10,
+        ];
+    }
+}